Added ensure newline and some other minor fixes
Re-ordered some things so that all of the pure text representation ans shuffeling is at the top, while the inner workings are more towards the bottom. Also removed quotes around the true flag in the word_wrap field, seems to have snuck in because the default, "auto", needs to be with quotes.